requirements.txt
streamlit=0.87.0
opencv=4.0.1
pillow=8.0.1
numpy=1.20.2
pybase64=1.2.0